DDCNL-9071: Remove IncomeCategoryHodFormatters
