+++ app/uk/gov/hmrc/tai/model/domain/calculation/TotalTax.scala
@@ -18,6 +18,8 @@ package uk.gov.hmrc.tai.model.domain.calculation
 
 import play.api.libs.json._
 import uk.gov.hmrc.tai.model.domain.taxAdjustments.TaxAdjustment
+import play.api.Logging
+import uk.gov.hmrc.tai.model.domain.calculation.TaxBand.taxBandReads
 
 case class TaxBand(
   bandType: String,
@@ -29,8 +31,34 @@ case class TaxBand(
   rate: BigDecimal
 )
 
-object TaxBand {
+object TaxBand extends Logging {
   implicit val formats: OFormat[TaxBand] = Json.format[TaxBand]
+
+  val taxBandReads = new Reads[Option[TaxBand]] {
+    override def reads(json: JsValue): JsResult[Option[TaxBand]] = {
+      val bandType = (json \ "bandType").as[String]
+      val code = (json \ "taxCode").as[String]
+      val income = (json \ "income").asOpt[BigDecimal]
+      val tax = (json \ "tax").asOpt[BigDecimal]
+      val lowerBand = (json \ "lowerBand").asOpt[BigDecimal]
+      val upperBand = (json \ "upperBand").asOpt[BigDecimal]
+      val rate = (json \ "rate").as[BigDecimal]
+      (income, tax) match {
+        case (Some(income), Some(tax)) =>
+          JsSuccess(Some(TaxBand(bandType, code, income, tax, lowerBand, upperBand, rate)))
+        case (None, None) =>
+          logger.info("Empty tax returned, no income or tax")
+          JsSuccess(None)
+        case (Some(income), None) =>
+          logger.error(s"Income value was present but tax was not in tax band: $bandType, code: $code")
+          JsSuccess(Some(TaxBand(bandType, code, income, 0, lowerBand, upperBand, rate)))
+        case (None, Some(_)) =>
+          val x = new RuntimeException(s"Tax value was present at income was not in tax band: $bandType, code: $code")
+          logger.error(x.getMessage, x)
+          JsSuccess(None)
+      }
+    }
+  }
 }
 
 sealed trait IncomeCategoryType
@@ -58,6 +86,47 @@ case class IncomeCategory(
 
 object IncomeCategory {
   implicit val formats: OFormat[IncomeCategory] = Json.format[IncomeCategory]
+
+  val incomeCategorySeqReads = new Reads[Seq[IncomeCategory]] {
+    override def reads(json: JsValue): JsResult[Seq[IncomeCategory]] = {
+      val categoryNames =
+        Seq("nonSavings", "untaxedInterest", "bankInterest", "ukDividends", "foreignInterest", "foreignDividends")
+      val incomeCategoryList = incomeCategories(json, categoryNames)
+      JsSuccess(incomeCategoryList)
+    }
+  }
+
+  private def categoryTypeFactory(category: String) =
+    category match {
+      case "nonSavings"       => NonSavingsIncomeCategory
+      case "untaxedInterest"  => UntaxedInterestIncomeCategory
+      case "bankInterest"     => BankInterestIncomeCategory
+      case "ukDividends"      => UkDividendsIncomeCategory
+      case "foreignInterest"  => ForeignInterestIncomeCategory
+      case "foreignDividends" => ForeignDividendsIncomeCategory
+      case _                  => throw new RuntimeException("Wrong income category type")
+    }
+
+  private def incomeCategories(json: JsValue, categories: Seq[String]): Seq[IncomeCategory] = {
+    val categoryMap: Map[String, Option[JsObject]] =
+      categories.map(category => (category, (json \ "totalLiability" \ category).asOpt[JsObject])).toMap
+
+    categoryMap.collect { case (category, Some(jsObject)) =>
+      val totalTax = (jsObject \ "totalTax").asOpt[BigDecimal].getOrElse(BigDecimal(0))
+      val totalTaxableIncome = (jsObject \ "totalTaxableIncome").asOpt[BigDecimal].getOrElse(BigDecimal(0))
+      val totalIncome = (jsObject \ "totalIncome" \ "amount").asOpt[BigDecimal].getOrElse(BigDecimal(0))
+      val taxBands =
+        (jsObject \ "taxBands").asOpt[Seq[TaxBand]](Reads.seq[Option[TaxBand]](taxBandReads).map(_.flatten))
+      val inComeCategory = categoryTypeFactory(category)
+      IncomeCategory(
+        inComeCategory,
+        totalTax,
+        totalTaxableIncome,
+        totalIncome,
+        taxBands.getOrElse(Seq()).filter(_.income > 0)
+      )
+    }.toList
+  }
 }
 
 case class TotalTax(
