+++ test/uk/gov/hmrc/tai/service/TotalTaxServiceSpec.scala
@@ -20,7 +20,7 @@ import org.mockito.ArgumentMatchers.{any, eq => meq}
 import play.api.libs.json.{JsObject, Json}
 import uk.gov.hmrc.tai.connectors.TaxAccountConnector
 import uk.gov.hmrc.tai.model.domain.calculation.IncomeCategory
-import uk.gov.hmrc.tai.model.domain.formatters.IncomeCategoryHodFormatters
+import uk.gov.hmrc.tai.model.domain.calculation.IncomeCategory.incomeCategorySeqReads
 import uk.gov.hmrc.tai.model.domain.taxAdjustments._
 import uk.gov.hmrc.tai.model.tai.TaxYear
 import uk.gov.hmrc.tai.repositories.deprecated.TaxAccountSummaryRepository
@@ -31,7 +31,7 @@ import scala.concurrent.Future
 class TotalTaxServiceSpec extends BaseSpec {
   val mockTaxAccountConnector: TaxAccountConnector = mock[TaxAccountConnector]
   val mockTaxAccountSummaryRepository: TaxAccountSummaryRepository = mock[TaxAccountSummaryRepository]
-  class Dummy extends IncomeCategoryHodFormatters
+  class Dummy
   val incomeCategoryHodFormatters = new Dummy
 
   override protected def beforeEach(): Unit = {
@@ -102,7 +102,7 @@ class TotalTaxServiceSpec extends BaseSpec {
       val result = sut.totalTax(nino, TaxYear()).futureValue
 
       result.incomeCategories must contain theSameElementsAs incomeCategories.as[Seq[IncomeCategory]](
-        incomeCategoryHodFormatters.incomeCategorySeqReads
+        incomeCategorySeqReads
       )
       result.reliefsGivingBackTax mustBe None
       result.otherTaxDue mustBe None
